Change navigation of Dynamic link click from  MeasureNow page to EnterDetails page
