+++ app/src/main/AndroidManifest.xml
@@ -158,15 +158,8 @@
             android:name=".ui.measure.EnterDetails"
             android:exported="true"
             android:screenOrientation="portrait"
-            android:windowSoftInputMode="stateHidden|adjustPan" />
-        <activity
-            android:name=".ui.measure.MeasureNow"
-            android:exported="true"
             tools:ignore="AppLinkUrlError"
-            android:screenOrientation="portrait">
-            <meta-data
-                android:name="android.app.lib_name"
-                android:value="" />
+            android:windowSoftInputMode="stateHidden|adjustPan" >
             <tools:validation testUrl="https://hhcustomtailor.com/measurement" />
             <intent-filter android:autoVerify="true">
                 <action android:name="android.intent.action.VIEW"/>
@@ -176,6 +169,15 @@
                 <data android:scheme="http"/>
                 <data android:scheme="https"/>
             </intent-filter>
+        </activity>
+        <activity
+            android:name=".ui.measure.MeasureNow"
+            android:exported="false"
+            android:screenOrientation="portrait">
+            <meta-data
+                android:name="android.app.lib_name"
+                android:value="" />
+
         </activity>
         <activity
             android:name=".ui.customization.Fabric_Select.order.OrderCartAct"
