+++ app/src/main/java/com/ms/hht/ui/measure/EnterDetails.java
@@ -1,16 +1,38 @@
 package com.ms.hht.ui.measure;
 
+import android.app.Dialog;
 import android.content.Intent;
 import android.graphics.Color;
+import android.graphics.drawable.ColorDrawable;
+import android.net.Uri;
 import android.os.Bundle;
+import android.text.TextUtils;
+import android.util.Log;
+import android.view.Gravity;
 import android.view.View;
+import android.view.Window;
+import android.widget.Button;
+import android.widget.ImageView;
+import android.widget.LinearLayout;
+import android.widget.TextView;
 
+import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 
+import com.google.android.gms.tasks.OnFailureListener;
+import com.google.android.gms.tasks.OnSuccessListener;
+import com.google.firebase.FirebaseApp;
+import com.google.firebase.dynamiclinks.FirebaseDynamicLinks;
+import com.google.firebase.dynamiclinks.PendingDynamicLinkData;
 import com.ms.hht.R;
+import com.ms.hht.data.SignUpResponse;
+import com.ms.hht.data.request.SignupRequest;
+import com.ms.hht.data.service.APICallList;
+import com.ms.hht.data.service.DisposableData;
 import com.ms.hht.databinding.ActEnterDetailsBinding;
 import com.ms.hht.utils.ComUserProfileData;
 import com.ms.hht.utils.CommFunc;
+import com.ms.hht.utils.InternetConnection;
 import com.ms.hht.utils.SessionManager;
 
 public class EnterDetails extends AppCompatActivity implements View.OnClickListener {
@@ -23,6 +45,8 @@ public class EnterDetails extends AppCompatActivity implements View.OnClickListe
     String genderForMeas = "Male";
     int HeightData, WeightData;
     SessionManager sessionManager;
+    SignUpResponse signUpResponse;
+    String userToken;
 
 
     @Override
@@ -47,6 +71,46 @@ public class EnterDetails extends AppCompatActivity implements View.OnClickListe
         detailsBinding.LooseBtn.setOnClickListener(this);
         detailsBinding.maleBtnTapped.setOnClickListener(this);
         detailsBinding.femaleBtnTapped.setOnClickListener(this);
+        Intent appLinkIntent = getIntent();
+        String appLinkAction = appLinkIntent.getAction();
+        Uri appLinkData = appLinkIntent.getData();
+        if(appLinkData!=null) { // Intent received from click on URL or scan URL
+            Log.d("keyss..", "Received Data here==>activion=" + appLinkAction);
+            Log.d("keyss..", "Received Data here==>appLinkData=" + appLinkData);
+            try {
+                FirebaseApp.initializeApp(getApplicationContext());
+            } catch (Exception e) {
+                e.printStackTrace();
+            } finally {
+                try {
+                    FirebaseDynamicLinks.getInstance()
+                            .getDynamicLink(appLinkIntent)
+                            .addOnSuccessListener(this, new OnSuccessListener<PendingDynamicLinkData>() {
+                                @Override
+                                public void onSuccess(PendingDynamicLinkData pendingDynamicLinkData) {
+                                    // Get deep link from result (may be null if no link is found)
+                                    Uri deepLink = null;
+                                    String token = null;
+                                    if (pendingDynamicLinkData != null) {
+                                        deepLink = pendingDynamicLinkData.getLink();
+                                        token = deepLink.getQueryParameter("token");
+                                        userLoginByToken(token);
+
+                                    }
+                                    Log.w("keysss....Success", "Receied Link data=>" + "__token==>" + token);
+                                }
+                            })
+                            .addOnFailureListener(this, new OnFailureListener() {
+                                @Override
+                                public void onFailure(@NonNull Exception e) {
+                                    Log.w("keysss....Error", "getDynamicLink:onFailure", e);
+                                }
+                            });
+                } catch (Exception e) {
+                    e.printStackTrace();
+                }
+            }
+        }
     }
 
     @Override
@@ -242,6 +306,116 @@ public class EnterDetails extends AppCompatActivity implements View.OnClickListe
         startActivity(in);
         finish();
         overridePendingTransition(R.anim.slide_in_right, R.anim.slide_out_left);
+    }
+    private void userLoginByToken(String token) {
+        if (token != null && !TextUtils.isEmpty(token))
+            if (InternetConnection.isConnected(EnterDetails.this)) {
+                CommFunc.ShowProgressbar(this);
+                Log.d("keyss...","Trying to login==>"+token);
+                this.userToken =  token;
+                APICallList.userLoginByToken(token, "user login", response, EnterDetails.this);
+            } else {
+                CommFunc.ShowStatusPop(this, getResources().getString(R.string.internet), false);
+            }
+    }
+    private void userLogin(String email, String password, String token) {
+        if (password != null && !TextUtils.isEmpty(password) && token != null && !TextUtils.isEmpty(token)) {
+            if (InternetConnection.isConnected(EnterDetails.this)) {
+                CommFunc.ShowProgressbar(this);
+                Log.d("keyss...", "Trying to login==>" + token);
+                SignupRequest signupRequest = new SignupRequest(email, password);
+                signupRequest.setToken(token);
+                APICallList.userLogin(signupRequest, "user login via passcode", response, EnterDetails.this);
+            } else {
+                CommFunc.ShowStatusPop(this, getResources().getString(R.string.internet), false);
+            }
+        }else {
+            CommFunc.ShowStatusPop(this, "Request could not be processed, require more information ", false);
+        }
+    }
+
+    private final DisposableData response = new DisposableData() {
+        @Override
+        public void onSuccess(String url_type, Object o) throws Exception {
+            if(url_type.equalsIgnoreCase("user login")){
+                CommFunc.DismissDialog();
+                signUpResponse = (SignUpResponse) o;
+                if (signUpResponse != null) {
+                    if(signUpResponse.getCode()==1) {
+//                        sessionManager = new SessionManager(EnterDetails.this);
+                        if(signUpResponse.getData().getEmail()!=null && signUpResponse.getData().getPassword()!=null) {
+                            sessionManager.logoutUser();
+                            sessionManager.setUserEmail(signUpResponse.getData().getEmail());
+                            sessionManager.setUserPassword(signUpResponse.getData().getPassword());
+                            sessionManager.setUserToken(userToken);
+                            userLogin(signUpResponse.getData().getEmail(), signUpResponse.getData().getPassword(), EnterDetails.this.userToken);
+                        }else{
+                            CommFunc.ShowStatusPop(EnterDetails.this, "Request could not be processed, requested information not found", false);
+                        }
+//                        sessionManager.setUserId("349");//For Testing need to be changed to signUpResponse.getData().getId());
+//                        sessionManager.setLoginSession("Logged in");
+//                        ShowPopUP("Logged in as "+signUpResponse.getData().getEmail(),true,signUpResponse);
+                    }
+                    else if (signUpResponse.getCode() == 2){
+                        ShowPopUP(signUpResponse.getMessage(),false,signUpResponse);
+                    }
+                    else {
+                        CommFunc.ShowStatusPop(EnterDetails.this,signUpResponse.getMessage(),false);
+                    }
+                }
+            }else if(url_type.equalsIgnoreCase("user login via passcode")){
+                CommFunc.DismissDialog();
+                signUpResponse = (SignUpResponse) o;
+                if (signUpResponse != null) {
+                    if(signUpResponse.getCode()==1) {
+                        sessionManager = new SessionManager(EnterDetails.this);
+                        Log.d("keyss....=>","id==>"+String.valueOf(signUpResponse.getData().getId()));
+                        sessionManager.setUserId(String.valueOf(signUpResponse.getData().getId()));
+                        sessionManager.setLoginSession("Logged in");
+                        ShowPopUP("Logged in as "+sessionManager.getUserEmail(),true,signUpResponse);
+                    }
+                    else if (signUpResponse.getCode() == 2){
+                        ShowPopUP(signUpResponse.getMessage(),false,signUpResponse);
+                    }
+                    else {
+                        CommFunc.ShowStatusPop(EnterDetails.this,signUpResponse.getMessage(),false);
+                    }
+                }
+            }
+        }
+
+        @Override
+        public void onError(String message) throws Exception {
+            CommFunc.DismissDialog();
+            CommFunc.ShowStatusPop(EnterDetails.this, message, false);
+        }
+    };
+    private void ShowPopUP(String message, boolean b, SignUpResponse signUpResponse) {
+        final Dialog dialog = new Dialog(EnterDetails.this);
+        Window window = dialog.getWindow();
+        assert window != null;
+        window.setGravity(Gravity.CENTER);
+        window.setLayout(LinearLayout.LayoutParams.WRAP_CONTENT, LinearLayout.LayoutParams.WRAP_CONTENT);
+        dialog.setContentView(R.layout.pop_up_status);
+        dialog.setCancelable(false);
+        dialog.getWindow().setBackgroundDrawable(new ColorDrawable(Color.TRANSPARENT));
+        dialog.show();
+
+        ImageView popheaderImage = dialog.findViewById(R.id.successPop_Header_image);
+        TextView poperrorDescription = dialog.findViewById(R.id.successPop_error_description);
+        Button submit = dialog.findViewById(R.id.successPop_done_button);
+
+        poperrorDescription.setText(message);
+
+        if (!b) {
+            popheaderImage.setImageResource(R.drawable.oopsimage);
+        } else {
+            popheaderImage.setImageResource(R.drawable.ok);
+        }
+        submit.setOnClickListener(v -> {
+            dialog.dismiss();
+
+        });
     }
 
 }
